requirements.txt
torch>=1.8.2
numpy>=1.19.2
Pillow>=8.0.0
scikit-image>=0.17.0
torchvision>=0.9.2
